Complete support for Currency items. Thank you Saerothan for fixing it.

+++ Trunk/src/hearthstone-world/Item.h
@@ -134,10 +134,10 @@ public:
 	
 	//! DB Serialization
 	void LoadFromDB( Field *fields, Player* plr, bool light );
-	void SaveToDB( int8 containerslot, int8 slot, bool firstsave, QueryBuffer* buf );
+	void SaveToDB( int16 containerslot, int16 slot, bool firstsave, QueryBuffer* buf );
 	bool LoadAuctionItemFromDB( uint64 guid );
 	void DeleteFromDB();
-    bool IsEligibleForRefund();
+	bool IsEligibleForRefund();
 	
 	HEARTHSTONE_INLINE void SoulBind()
 	{
